Refactor: Move RegexValidator factory methods to RegexValidatorFactory in core
